Create Download-Link
In this code, the /download/<filename> route is used to serve the generated PowerPoint file for download. When the user clicks the download link, the send_file function sends the file as an attachment to the user's browser, triggering the browser's download behavior.

Make sure to import the send_file function from the Flask library:

from flask import Flask, request, jsonify, send_file

Additionally, you should provide a link to the download route in your frontend interface once the conversion is complete. For example, you might display a link to the user like this:

<a href="/download/{{ ppt_filename }}" download>Click here to download the PowerPoint presentation</a>